+++ runners/native/installer.nsi
@@ -1,5 +1,5 @@
 ; =============================================================================
-; NSIS Installer Script for "{{project.name}}"
+; NSIS Installer Script for "{{name}}"
 ;
 ; Features:
 ; - Self-contained and robust function to add the installation directory to the PATH.
@@ -14,10 +14,10 @@
 
 ;--------------------------------
 ; General
-Name "{{project.name}}"
-OutFile ".\{{project.name}}-{{project.version}}-installer.exe"
-InstallDir "$APPDATA\gama\{{author.name}}\{{project.name}}"
-InstallDirRegKey HKCU "Software\{{author.name}}\{{project.name}}" "InstallDir"
+Name "{{name}}"
+OutFile ".\{{name}}-{{version}}-installer.exe"
+InstallDir "$APPDATA\gama\{{author.name}}\{{name}}"
+InstallDirRegKey HKCU "Software\{{author.name}}\{{name}}" "InstallDir"
 RequestExecutionLevel user
 
 
@@ -40,27 +40,27 @@ RequestExecutionLevel user
 ;=============================================================================
 ; Installer Section
 ;=============================================================================
-Section "Install {{project.name}}" SecInstall
+Section "Install {{name}}" SecInstall
   SetOutPath "$INSTDIR"
 
   ; --- Install application files ---
-  File "..\native\{{project.name}}.exe"
+  File "..\native\{{name}}.exe"
   File "..\native\libvgama.dll"
   File /r "..\..\assets"
 
   ; --- Create Uninstaller and Registry entries for Add/Remove Programs ---
-  WriteRegStr HKCU "Software\{{author.name}}\{{project.name}}" "InstallDir" "$INSTDIR"
+  WriteRegStr HKCU "Software\{{author.name}}\{{name}}" "InstallDir" "$INSTDIR"
   WriteUninstaller "$INSTDIR\uninstall.exe"
 
-  WriteRegStr HKCU "Software\Microsoft\Windows\CurrentVersion\Uninstall\{{author.name}}\{{project.name}}" "DisplayName" "{{project.name}}"
-  WriteRegStr HKCU "Software\Microsoft\Windows\CurrentVersion\Uninstall\{{author.name}}\{{project.name}}" "UninstallString" '"$INSTDIR\uninstall.exe"'
-  WriteRegDWORD HKCU "Software\Microsoft\Windows\CurrentVersion\Uninstall\{{author.name}}\{{project.name}}" "NoModify" 1
-  WriteRegDWORD HKCU "Software\Microsoft\Windows\CurrentVersion\Uninstall\{{author.name}}\{{project.name}}" "NoRepair" 1
+  WriteRegStr HKCU "Software\Microsoft\Windows\CurrentVersion\Uninstall\{{author.name}}\{{name}}" "DisplayName" "{{name}}"
+  WriteRegStr HKCU "Software\Microsoft\Windows\CurrentVersion\Uninstall\{{author.name}}\{{name}}" "UninstallString" '"$INSTDIR\uninstall.exe"'
+  WriteRegDWORD HKCU "Software\Microsoft\Windows\CurrentVersion\Uninstall\{{author.name}}\{{name}}" "NoModify" 1
+  WriteRegDWORD HKCU "Software\Microsoft\Windows\CurrentVersion\Uninstall\{{author.name}}\{{name}}" "NoRepair" 1
 
   ; --- Create shortcuts ---
   CreateDirectory "$SMPROGRAMS\{{author.name}}"
-  CreateShortCut "$SMPROGRAMS\{{author.name}}\{{project.name}}.lnk" "$INSTDIR\{{project.name}}.exe"
-  CreateShortCut "$DESKTOP\{{project.name}}.lnk" "$INSTDIR\{{project.name}}.exe"
+  CreateShortCut "$SMPROGRAMS\{{author.name}}\{{name}}.lnk" "$INSTDIR\{{name}}.exe"
+  CreateShortCut "$DESKTOP\{{name}}.lnk" "$INSTDIR\{{name}}.exe"
 SectionEnd
 
 ;=============================================================================
@@ -68,13 +68,13 @@ SectionEnd
 ;=============================================================================
 Section "Uninstall"
   ; --- Remove shortcuts and Start Menu folder ---
-  Delete "$SMPROGRAMS\{{author.name}}\{{project.name}}.lnk"
-  Delete "$DESKTOP\{{project.name}}.lnk"
+  Delete "$SMPROGRAMS\{{author.name}}\{{name}}.lnk"
+  Delete "$DESKTOP\{{name}}.lnk"
   ; RMDir "$SMPROGRAMS\{{author.name}}"
 
   ; --- Remove registry keys ---
-  DeleteRegKey HKCU "Software\Microsoft\Windows\CurrentVersion\Uninstall\{{author.name}}\{{project.name}}"
-  DeleteRegKey HKCU "Software\{{author.name}}\{{project.name}}"
+  DeleteRegKey HKCU "Software\Microsoft\Windows\CurrentVersion\Uninstall\{{author.name}}\{{name}}"
+  DeleteRegKey HKCU "Software\{{author.name}}\{{name}}"
 
   ; --- Remove all installed files and directories ---
   RMDir /r "$INSTDIR"
