+++ vgama/package.v
@@ -16,14 +16,8 @@ pub fn (p Project) package_native(inst Installation) ! {
 	mut template := os.read_file(template_path) or {
 		return error('Failed to read NSIS template at ${template_path}: ${err}')
 	}
-
-	// 3. Perform string replacements
-	template = template.replace('{{project.name}}', conf.name)
-	template = template.replace('{{project.version}}', conf.version.str())
-	template = template.replace('{{author.name}}', conf.author)
-
 	// 4. Write final NSIS script
-	os.write_file(nsi_script_path, template)!
+	os.write_file(nsi_script_path, conf.substitute(template))!
 
 	// 5. Execute makensis
 	nsis := $if windows { 'nsis' } $else { 'makensis' }
